Move config.rs in halley-wl to new crate halley-config
