+++ crates/halley-wl/src/lib.rs
@@ -1,7 +1,6 @@
 pub mod activity;
 pub mod anim;
 pub(crate) mod backend_iface;
-pub mod config;
 pub(crate) mod input;
 pub(crate) mod interaction;
 pub mod render;
